Move & rename intro content
